flexed class.h into GPU mode

+++ src/class_device.hpp
@@ -1,5 +1,5 @@
 #pragma once
-typedef float real;
+typedef double real;
 
 namespace Dens{
 	struct EpiDev{
